+++ package.json
@@ -56,10 +56,11 @@
 		"add-module-exports-webpack-plugin": "^0.1.0",
 		"ava": "^1.0.0-rc.1",
 		"awesome-typescript-loader": "^5.2.1",
-		"callsites": "^2.0.0",
+		"callsites": "^3.0.0",
 		"codecov": "^3.1.0",
 		"del-cli": "^1.1.0",
 		"dot-prop": "^4.2.0",
+		"is-node": "^1.0.2",
 		"is-valid-identifier": "^2.0.0",
 		"license-webpack-plugin": "^2.0.2",
 		"lodash.isequal": "^4.5.0",
