+++ source/types.d.ts
@@ -1,2 +1,2 @@
-declare module 'callsites';
 declare module 'is-valid-identifier';
+declare module 'is-node';
